+++ frontend/src/redux/store.js
@@ -1,6 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit'
 import cartReducer from '../redux/features/cart/cartSlice'
-import booksApi from './features/books/booksApi'
 import ordersApi from './features/orders/ordersApi'
 import authReducer from './features/auth/authStore'
 import googleBooksReducer from './features/googleBooks/googleBooksSlice'
@@ -10,9 +9,8 @@ export default configureStore({
     cart: cartReducer,
     auth: authReducer,
     googleBooks: googleBooksReducer,
-    [booksApi.reducerPath]: booksApi.reducer,
     [ordersApi.reducerPath]: ordersApi.reducer
   },
   middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware().concat(booksApi.middleware, ordersApi.middleware)
+    getDefaultMiddleware().concat(ordersApi.middleware)
 })