+++ frontend/src/routers/router.jsx
@@ -8,14 +8,6 @@ import { CheckoutPage } from "../pages/books/CheckoutPage";
 import { SingleBook } from "../pages/books/SingleBook";
 import { PrivateRoute } from "./PrivateRoute";
 import { OrderPage } from "../pages/books/OrderPage";
-import { AdminRoute } from "./AdminRoute";
-import { AdminLogin } from "../components/AdminLogin";
-import { DashboardLayout } from "../pages/dashboard/DashboardLayout";
-import { Dashboard } from "../pages/dashboard/Dashboard";
-import {ManageBooks} from "../pages/dashboard/manageBooks/ManageBooks";
-import AddBook from "../pages/dashboard/addBook/AddBook";
-import { UpdateBook } from "../pages/dashboard/editBook/UpdateBook";
-
 
 const router = createBrowserRouter([
     {
@@ -56,40 +48,6 @@ const router = createBrowserRouter([
             }
         ]
     },
-    {
-        path: "/admin",
-        element: <AdminLogin/>
-    },
-    {
-        path: "/dashboard",
-        element: <AdminRoute>
-            <DashboardLayout/>
-        </AdminRoute>,
-        children: [
-            {
-                path: "",
-                element: <AdminRoute><Dashboard/></AdminRoute>
-            },
-            {
-                path: "add-new-book",
-                element: <AdminRoute>
-                        <AddBook/>
-                    </AdminRoute>
-            },
-            {
-                path: "edit-book/:id",
-                element: <AdminRoute>
-                    <UpdateBook/>
-                </AdminRoute>
-            },
-            {
-                path: "manage-books",
-                element: <AdminRoute>
-                    <ManageBooks/>
-                </AdminRoute>
-            }
-        ]
-    }
 ])
 
 export default router;